Move `play_game_task` to validation with server
